\name{lmi}
\alias{lmi}
\title{Support of HAC methods within glm regression model}
\usage{
  lmi(formula, data, vcov = NULL, subset, weights,
    na.action, method = "qr", model = TRUE, x = FALSE,
    y = FALSE, qr = TRUE, singular.ok = TRUE,
    contrasts = NULL, offset, ...)
}
\description{
  Support of HAC methods within glm regression model
}
\seealso{
  \code{\link{lm}
}
}

