\name{QP.Norm}
\alias{QP.Norm}
\title{QP function for calculation of Sharpe Ratio}
\usage{
  QP.Norm(R, tau, scale = NA)
}
\arguments{
  \item{R}{an xts, vector, matrix, data frame, timeSeries
  or zoo object of asset returns}

  \item{tau}{Time Scale Translations Factor}

  \item{scale}{number of periods in a year (daily scale =
  252, monthly scale =}
}
\description{
  QP function for calculation of Sharpe Ratio
}
\seealso{
  \code{\link{CalmarRatio.Norm}}, \cr
}

