\name{glmi}
\alias{glmi}
\title{Support of HAC methods within lm regression model}
\usage{
  glmi(formula, family = gaussian, data, vcov = NULL,
    weights, subset, na.action, start = NULL, etastart,
    mustart, offset, control = list(...), model = TRUE,
    method = "glm.fit", x = FALSE, y = TRUE,
    contrasts = NULL, ...)
}
\description{
  Support of HAC methods within lm regression model
}
\seealso{
  \code{\link{glm}
}
}

